requirements.txt
pydantic-settings
pydantic
google-cloud-aiplatform[adk,agent_engine]
google-adk
jsonschema
